Fix typed locals for non-trivial lhs
Summary:
If the lhs of an assignment is a `list` or array access `$x[e]`, then we might need to do some enforcement, but we won't be able to do it as an `as` on the rhs, because we don't know the type to enforce. Instead, we add an `as` expression (or expression-statement depending on the enclosing context), to do the enforcement.

For example, if we have `list($x, $y) = e` and `$x` and `$y` are typed locals of type `int`, then  `e` could be a tuple of ints, or a vec of ints, so we instead just enforce that $x and $y are ints after the assignment.

If we have `$x[e1] = e2` and if `$x` is an enforced local of shape type, then we might not know what is the correct type of `e2` for the assignment, because the shape type could be in another decl behind a type alias or constant. Alternately, in FIXME'd code, we might not know what `e1` evaluates to. It would be possible to do an `as` on the rhs where we know the type of `e1`, but we leave that to the future.

The only way to observe the difference in doing the enforcement on the rhs vs in a statement after is to catch the exception when the enforcement fails and observe the variable being assigned in a different state.

Reviewed By: andrewjkennedy

Differential Revision: D47755623

fbshipit-source-id: fc81fb6c650db7d71a4fc13304d7bf51c78e7f4d