Add product attribution ID embed functions for propagating through eg. PSP
Summary:
Similarly to implicit context, we need the ability to propagate product IDs from closure creators into PSP and other deferred execution constructs in www.

The added 2 php implementations use get_product_attribution_id_internal to fetch the current attribution, whether it's an int, null, or lambda, and capture the value in a closure. We then set the attribution inside the closure before calling the original closure.

This makes it so that when we call get_product_attribution_id inside the original closure (eg. when calling out to things via thrift) we get the original attribution from the site where the closure was created.

I'm implementing this in standard library per oulgen's request to prevent userland usage of `get_product_attribution_id_internal`.

Reviewed By: jano

Differential Revision: D47674625

fbshipit-source-id: d6375d6c20f4e8db1b348581e7b02a9e3f6fb0bf