Refer to enums using the typedef they are being accessed through
Summary:
When a Thrift enum is used via a typedef within an `impl Default` or a `const`, the underlying type of the enum is not always directly accessible in the current crate. It may be behind a transitive dependency.

This diff fixes errors that look like:

```
error[E0433]: failed to resolve: use of undeclared crate or module `deployment`
    --> data_preproc/if/lib.rs:3646:26
     |
3646 |             releaseTier: deployment::types::MasterServiceTier::UNKNOWN,
     |                          ^^^^^^^^^^ use of undeclared crate or module `deployment`

error[E0433]: failed to resolve: use of undeclared crate or module `deployment`
    --> data_preproc/if/lib.rs:3778:54
     |
3778 |             releaseTier: field_releaseTier.unwrap_or(deployment::types::MasterServiceTier::UNKNOWN),
     |                                                      ^^^^^^^^^^ use of undeclared crate or module `deployment`
```

where "data_preproc/if/lib.rs" is really supposed to contain `ReleaseTier::UNKNOWN`, but the definition of ReleaseTier is something like `typedef deployment.MasterServiceTier ReleaseTier`.

Reviewed By: zertosh

Differential Revision: D47941012

fbshipit-source-id: 3848cdd791599286ca1381a902e8bbddc814f986