Recover exhaustivity check in the presence of intersected like types
Summary:
When an enum is pessimised we get a type of the form `bound & ~E` where `bound` is the visible bound of `E`.

Because exhaustivity check is partial when it comes to intersections, `bound` part of the intersection can turn of exhaustivity checking which turns of useful signal on enums.

This diff recovers it by exempting specifically this pattern.

Reviewed By: andrewjkennedy

Differential Revision: D47890092

fbshipit-source-id: 0f9a2978a0216afc9b7ca47042a83f57852888cc