is_complete is computed within glean.ml
Summary:
It felt dodgy to compute "is_completed" only on the final output of SymbolIndex. For instance if we requested `max_results:100` and glean delivered an incomplete list of the first 100 results, but five of them were tombstoned inside SymbolIndex, then we'd wrongly give the answer that autocomplete was complete.

In this diff I changed it so the bottom-level glean query is what decides if it's complete or not.

Reviewed By: mheiber

Differential Revision: D47915259

fbshipit-source-id: fc2ef732d6bc5ce3c587561d5f265d584756976f