mock sie_env
Summary:
I want hh_single_complete to take a .php file and report what query_text, context and kind_filter it derived from the AUTO332 and the namespace-map.

The cleanest way I can think is to create a new search_provider (in addition to NoIndex, LocalIndex, CustomIndex). This new one would be MockIndex and would just be a normal mock object...

Plan: when autocompleteService calls into our mocked search_provider, we'll learn what parameters it called into it with (which query_text, context, and kind_filter) and we'll be able to display them.

Reviewed By: mheiber

Differential Revision: D47928470

fbshipit-source-id: 83b06de0b170f4ecf98a865d339b7424f9591a86