Avoid proxying certain FS operations to the CLI client
Summary:
Usually CLI clients run code out of a repository that is also used to serve `http` requests. In these instances read operations against files inside the repository should succeed un-proxied as apache should also have read access to the repository. This diff adds an option to avoid proxying `stat()`, `lstat()` and `open()` (when only for reading) when the target file is in the repository. Additionally support for both not proxying `realpath()` and eliding it all together are now available (note that the path to the repository has already been resolved via realpath and paths within repositories don't generally involve symlinks).

As with the xbox changes in this stack new runtime options have been added to get this behavior.

Reviewed By: ricklavoie

Differential Revision: D47884561

fbshipit-source-id: b9e4084e25ea89203463d49ae3cdae3ff9f55935