Add xbox handling
Summary:
Introduce support for proxying xbox threads to a CLI client. The client will spin up worker threads to handle requests from xbox clients.

A runtime option controls this new behavior and is read from the client allowing this change to be both push safe and easily disabled for any incompatible clients on an individual basis.

Reviewed By: ricklavoie

Differential Revision: D47884560

fbshipit-source-id: 57103bdb75c7ce6a8f377f7a818ff9a8de1de8e0