hh_single_complete only decls the <?hh files
Summary:
Our hh_single_complete tests (for autocomplete, and shortly for search) take two kinds of input files in their multifile:
* <?hh files which define symbols so that the searches we do reflect those symboles
* files with newline-separated lists of query_texts

We might conceivably get rid of the "newline-separated" paradigm for autocomplete (since we can wholly switch over to AUTO332). But we won't get rid of it for search.

This diff makes sure we don't try to send the "newline-separated" files through decling.

Reviewed By: mheiber

Differential Revision: D47952728

fbshipit-source-id: 5069534e80cef0ee6ae142b681d814e93592731e