Add flag to keep user attributes to Hack server config
Summary:
Zoncolan misses a common pattern where attributes are declared on an interface methods and the methods' implementation do not have this attribute declared explicitly. This prevents DSL queries on attributes to match to methods which inherit attributes.

An example of this is GraphQL entrypoints are written using this pattern, where there is an interface with attributes declared on the methods, but the actual implementation of the methods have no explicit attributes. For this reason, it's not possible to write a model generator that creates models for GraphQL entrypoints using DSL.

[More details on issue + current workarounds](https://docs.google.com/document/d/1eA4Q5-TeAKLoG540qvIy6_4Qz5vwIh3hF9h-1ohacP0/edit?usp=sharing)
[More details on implementation](https://docs.google.com/document/d/1L7b5zb1E3F4t0dDeWmvhfBp9KWeYTvW4ogjbu3YUnXc/edit?usp=sharing)

# This diff
Adds a flag to the hack server config so that Zonk can be configured to keep all attributes, which will then be used to compute inherited attributes on methods

See D47775649, where flag is added to hack config for Zonk runs

Reviewed By: vsiles

Differential Revision: D47773914

fbshipit-source-id: 92f2b94b2408195cc989593ccc6718d6d6abdf9c