Fix memory effects of $this in debug_backtrace(), disable it without EnableArgsInBacktraces
Summary:
Do not report `$this` in debug_backtrace() without EnableArgsInBacktraces. It
was already not available for inlined frames, so its usefullness was limited,
especially in production configurations with heavy inlining.

Fix its memory effects when EnableArgsInBacktraces is enabled, so that we don't
optimize away ctx. This will make it possible to improve sandbox debugging
experience in the future by removing the restriction for inlined frames.

Also unblocks the next diff that would otherwise optimize ctx away.

Reviewed By: ricklavoie

Differential Revision: D47982566

fbshipit-source-id: 269038b2c20a11e0cf3dd67ab3c28d789acf2dbe