Add an elab pass to remove memoization attributes and plug it into textual translation
Summary:
The memoization effect is irrelevant when it comes to the kind of analysis we
do in Infer but its presence complicates the resulting IR and confuses the analyser.

This pass removes memoization attributes at the AST level before it reaches the
IR and allows us to get clean `textual` output.

Reviewed By: enetsee

Differential Revision: D47963526

fbshipit-source-id: b4acaa06885ad972ff05a84f501775a49a52e4a6