thrift-rust: compilation failure on optionals with defaults without deprecated_optional_with_default_is_some flag
Summary:
Thrift optional field with defaults should be initialized to `None` according to the standard. The current behavior is to initialize it to `Some`.

In order to transition to the new behavior (which is a breaking change), a thrift compiler flag `deprecated_optional_with_default_is_some` is introduced. In the first stage, it will be a compilation error to use optionals with defaults in rust without this flag. In a later change, using optionals with defaults without the flag will generate spec-conforming code.

Reviewed By: zertosh

Differential Revision: D47837746

fbshipit-source-id: 91a9c95a5d79c6b444cc0da41283382b7146c8ec