Change shape_type tuple to record
Summary:
This changes the shape_type tuple introduced in the parent diff to a record. The name of `s_unknown_value` represents the type of the values of `...` fields in a shape, i.e. the result of `Shapes::idx` on open shapes and `nothing` for closed shapes. The follow up diff will introduce `s_unknown_key`.

Notes / style choices:

- This diff is behavior preserving. I have noted inline several places where I think we made mistakes in the handling elements of the Tshape constructor, see `TODO(shapes)` comments
- I've preserved the names of intermediate variables to the point of extreme verbosity in pattern matching, otherwise this would be super painful to review.
  - We can migrate to field punning and record update syntax in follow up diffs
- In places where we only matched a single element of the tuple e.g. `Tshape (_, _, fdm)`, I used a record with a wildcard `Tshape { s_fields = fdm; _ }`.
- For places where we matched on 2 or 3 elements, even if they're threaded through unchanged, I destructured the entire record.

Differential Revision: D47781731

fbshipit-source-id: 4fe667e7f6f6d3c92fd532cf8722fca8180fac6c