Improve latency of TokenBucketConcurrencyController
Summary:
This diff fixes the p99 problem of TokenBucketConcurrencyController by adjusting it's algorithm to have different logic in fast mode and in slow mode.

In fast mode TBCC is now accounts for the fact that some percentage of dequeue() operations will fail. It now maintains it's own counter for the number of pending dequeue operations and retries dequeue operations till the counter is > 0.

In slow mode TBCC does not retry (as the queue is full) and treats dequeue() failure as an indicator that queue is empty at the moment. Slow mode is now cleared not by dequeue() returning nullopt but rather by pendingDequeueOps == 0.

Last change is that fast mode is not enabled concurrently with the slow mode.

Reviewed By: sethdelliott

Differential Revision: D48042556

fbshipit-source-id: 960c12260c57bd8d35b32490d964cba50b7e8069