move ServerCommandTypes.Find_refs => SearchTypes
Summary:
I wish to implement find-all-refs in glean. Therefore Glean.ml (the module which constructs glean angle queries) must be moved into a low-enough level module. I'm moving it into SearchTypes, which feels like the natural place.

This diff is an alternative to D47991228. I prefer this one, based on rjbailey's comments on the other.

Reviewed By: rjbailey

Differential Revision: D47998199

fbshipit-source-id: 9a8567691e19a3006d2b854c23165d0b9a68d17c