Improve stacktrace of asio cycle exceptions
Summary:
Asio maintains an invariant that the subgraph of the dependency graph that
contains only the first unfinished child does not contain a cycle. To maintain
this invariant, we perform online cycle detection in two situations:

- a resumed async function or generator awaits an Awaitable
- the first unfinished child of an await all or concurrent wait handle
  finishes, changing the first unfinished child

The first situation happens within the context of the async function or
generator, so a cycle exception will have proper stacktrace. However, in the
second case, the logic is executed in the context of the scheduler, resulting
in a truncated stacktrace.

Fix this by recomputing the stacktrace from the await all or concurrent wait
handle.

Reviewed By: ricklavoie

Differential Revision: D48085017

fbshipit-source-id: 3308a89c849be9d558e2ee9fd82049e0934ea9c3