Shouldn't define Flags at top-level in a header file
Summary: type-constraint.h was 'using' in a header file to make a shorthand for TypeConstraintFlags.  I suspect the intention was to put the 'using' inside the class - but really it's not a good idea to be defining a short name at such a broad scope anyway - the longer name is fine.

Reviewed By: ricklavoie

Differential Revision: D48085330

fbshipit-source-id: 78c653affde16d21bd550f60d3785d07788c9533