Unhack sink and stream more
Summary:
Sink and stream are not types in the usual sense but grammatical constructs that associate various types and exceptions with functions. D47840589 mostly decoupled them from the actual return type and this is a follow-up diff that cleans up AST hierarchy and visitors.

In particular, this diff removes the useless `t_templated_type` AST node and `add_type_instantiation_visitor`. The latter is redundant because we already have a container visitor.

Move the ast visitor test to tests, next to the closely related `ast-testing` target while at it. This way it is no longer excluded from `buck test thrift/compiler/test/...` which was a common source of missed signal when running tests locally.

Reviewed By: iahs

Differential Revision: D47959921

fbshipit-source-id: 52e1fc008b6861b056a5182b2dd2db0384ffcfbf