Rust Service: Arc<T> self-implement service trait if T does
Summary:
The generated code ensured that if `T` implements the service trait then `Box<T>` does as well.
This diff extends the same logic for `Arc<T>` as well.

The reason for the change is that if we have a service implementation struct containing data that needs to be alive for the whole duration of the application (thinking to a connection pool) then having the instance wrapped in `Arc` and then just "cloning" it makes the experience much smoother.

Differential Revision: D47993415

fbshipit-source-id: 16468902698d41e5949ee0c05d1575d2506823f4