Allow moving from AsyncSocket, transfer of FD read state from AsyncFdSocket
Summary: The current APIs are a bit clunky, and are intended only for usage in TLS code. If this keeps coming up, it is possible to get closer to real move semantics that works on both `AsyncFdSocket` (with pre-existing read and/or write state) and non-FD `AsyncSocket`. However, figuring out the correct details for the more general thing is considerably more involved, and is just not worth it yet.

Reviewed By: AjanthanAsogamoorthy

Differential Revision: D48087183

fbshipit-source-id: fb9cf0484733a16648f276fbb5873375d7a1f0c4