Use `AsyncFdSocket` for `AF_UNIX` connections
Summary:
The subsequent D48048366 uses this to close a race condition where FDs are sent by a client that completed the StopTLS handshake, before the server had managed to `moveToPlaintext`.

The gating to exclude Windows is here to avoid figuring out how to compile `SocketFds` (which uses `std::variant`) on older MSVC versions, which are still in use in some places.

Reviewed By: AjanthanAsogamoorthy

Differential Revision: D48087182

fbshipit-source-id: b2168283f8d9bfffe67ccf56c8e8eb89678bed6a