expose find_refs through SymbolIndex
Summary:
This diff is just plumbing. It exposes Glean.query_refs up through CustomSearchService and thence through SymbolIndex, which is the API that's consumed by the rest of the codebase.

(Next: I will be writing ClientIdeDaemon to invoke SymbolIndex.find_refs)

This diff is exciting! It's the first part of work specifically in the direction of fast find-refs!

Reviewed By: mheiber

Differential Revision: D48015389

fbshipit-source-id: 6f6db01feb768640f79b6cc777cf5c94412b5751