simplify findRefsService a little
Summary:
This is a small refactor. I think the code in findRefsService was needlessly complicated. Its job is, given a list of files, to find all occurrences of a target in their TASTs.

It used to first map each filename into a filename+TAST, then fold over those and add occurrences into an accumulator. (note: the function signature takes `acc` as an input to this batch of files, but in reality, the way we wrote MultiWorker is that `acc` always ends up being the empty list at the start of this function).

I changed it to just have a single loop.

My motive is I want to support streaming errors, i.e. print errors the moment they're discovered. I specifically don't want to delay for all TASTs before the first result is printed.

Reviewed By: nt591

Differential Revision: D48092444

fbshipit-source-id: be1a5df71631ebde499817580105bff1e6d5e10d