Handle the case when toThrift(...) is non-comparable
Summary:
The existing implementation assumes `toThrift(...)` always comparable, which is not true. This diff re-implemented some logic in Adapt.h so that it can be handled properly.

Eventually we will use `op::equal` to replace the codegen so that this part of the logic in `Adapt.h` can be removed.

Reviewed By: iahs

Differential Revision: D48059657

fbshipit-source-id: 018d49047a621fab1fea70a2a2a6a8f9c5c2e257