Make check_args tail-recursive to avoid stack overflow for large number of arguments
Summary:
The Hack type-checker segfaults when a function is passed a huge number of arguments (most easily done with a variadic function). This is due to a non-tail-recursive function `check_args` in the typing of calls. Solution: make it tail-recursive (somewhat clumsily - this logic could do with a bit of a clean-up).

Note that HackC (and possibly the runtime) choke on such a large number of arguments anyway, but we want to reuse the function calling logic for typing of `vec[...]`, for which HackC and the runtime do support a very large number of elements.

Reviewed By: vsiles

Differential Revision: D48174651

fbshipit-source-id: 06d5b523ac2a02b9e8ea4601980be5b2758783e5