Use expected type for lambda even if not a function type
Summary:
Collection constructors such as `vec` behave differently from ordinary functions with respect to bidirectional typing of lambdas. For example,
```
class C {
  public function foo():void { }
}
function expectVecFun(vec<(function(C):void)> $v):void { }

function test():void {
  expectVecFun(vec[$x ==> $x->foo()]);
}
```
is accepted by Hack, but replace the `vec[_]` with a call to `makeSingle` as below, and the code is rejected.
```
function makeSingle<T>(T $x):vec<T> {
  return vec[$x];
}
function test():void {
  expectVecFun(makeSingle($x ==> $x->foo()));
}

```
That's because (a) we don't propagate an "expected type" that's a type variable, even if has a useful upper bound (here, `#1 <: (function(C):void)` where `#1` is the fresh type variable generated for `T` in `makeSingle`, and the constraint comes from the check against the parameter type in `expectVecFun`), and (b) we don't use the expected type when checking a lambda if it's not an already-solved function type.

So to fix (a) we don't ignore type variables with upper bounds when setting the expected type, and (b) we deal with in the drop-through case for lambdas.

Differential Revision: D47993179

fbshipit-source-id: 212fca3cccde9ef547e0fbb2aad00039c6ad02fd