Store class constant inferred types more efficiently in HHBBC
Summary:
If we infer a type for a class constant, which isn't a
scalar, we store it in a global table. A global table is used to keep
memory usage to a minimum. While the table is a
folly_concurrent_hash_map_simd, which should (in theory) scale to high
insertions/sec, it still has become a bottleneck. I believe this is
due to increasing usage of enum classes, where the "constants" are
typically objects, which will never be scalars.

To scale this better, store a vector of class constant types in each
ClassInfo. This doesn't seem to use a meaningful amount of more
memory, but is a *lot* faster because each vector can be inserted into
completely independently.

Reviewed By: jano

Differential Revision: D48276281

fbshipit-source-id: 6802f46825322cd062213322ee91e05e3edb2479