Revert D48194024: Multisect successfully blamed "D48194024: Use Semifuture to implement Future thrift methods" for test or build failures
Summary:
This diff is reverting D48194024
D48194024: Use Semifuture to implement Future thrift methods by bmaurer has been identified to be causing the following test or build failures:

Tests affected:
- [thrift/lib/cpp2/test:ThriftClientTest - ThriftClientTest.FutureCallRequestResponse](https://www.internalfb.com/intern/test/562950035124227/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2760440
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: bmaurer

Differential Revision: D48292722

fbshipit-source-id: 4c1a0aaa782d87c2cea13cb5affc4f787c951d04