Add NoAutoLikes to Vector methods
Summary:
Add `<<__NoAutoLikes>>` to methods on `Vector<T>` that return a `Vector<T>` or related type. This is sound because elements of `Vector` are already pessimised: calling `get` or indexing a `Vector<t>` will return a `~t`. The `Vector` bit of the type is enforced by HHVM, so there is no need for an additional `~`.

We avoid annotating the higher-order methods as there are issues with completeness of lambda inference that introduce new errors into code.

Differential Revision: D48242833

fbshipit-source-id: 9400987f27c9ff1fff323f5fd78bc20ab4f8d755