frame parser that use allocats buffer the size of the frame
Summary:
`AllocatingParserStrategy` is an extension to `ParserStrategy` that allocates a buffer using an `Allocator`, default to be `std::allocator` rather than the `folly::IOBuf` chain based on the size of the frame. This strategy does not support `void readBufferAvailable(std::unique_ptr<folly::IOBuf>) `.

* `getReadBuffer(...)` returns the write head of current writable buffer.
* `readDataAvailable(len)` will increment the write head (`size_`) by `len`. handle all the pending rocket frame in the current buffer to `owner` and manage excess bytes.
* since `readBufferAvailable(...)` is not supported, we return `false` for `isBufferMovable(...)`

Performance wise, it will be a bit worse than the default parsing strategy, however, it will return a continuous buffer rather than a chained buffer, some application may prefer this behavior.

Reviewed By: robertroeser

Differential Revision: D48278286

fbshipit-source-id: 0fa18dbe324744181eb5a0e30104b0a8f2206bd0