Standardize on (string * Pos.absolute) in clientLsp too
Summary:
The standard within findRefsService and FindRefsWireFormat is that find-references returns `(string * Pos.absolute) list`.

Prior to this diff, clientLsp was filtering it to just Pos.absolute list. There's no need for that, and it makes things irregular (because both clientLsp.ml and findRefsService.ml will want to stream their partial results to the find-refs streaming file). So I'm normalizing clientLsp.

Reviewed By: nt591

Differential Revision: D48298170

fbshipit-source-id: b87d22faa65e09fdf52031568253dd6d6d3d7973