Gate setting headers into $_SERVER behind a flag
Summary:
The Secure Application Frameworks team has an initiative to reduce the surface area of request state in superglobals. As part of this, headers are meant to be requested raw via the init stack and handled in a protected manner rather than exposing via super globals (that then get unset).

In the future, it is the hope that this value may get passed down via the entrypoint directly, but that is for a later time.

Reviewed By: ricklavoie

Differential Revision: D48318659

fbshipit-source-id: cac8e0e218787af282e6e6ec784c11965217f358