Implement DSL matching for inherited attributes
Summary:
# Context
Zoncolan misses a common pattern where attributes are declared on an interface methods and the methods' implementation do not have this attribute declared explicitly. This prevents DSL queries on attributes to match to methods which inherit attributes.

An example of this is GraphQL entrypoints are written using this pattern, where there is an interface with attributes declared on the methods, but the actual implementation of the methods have no explicit attributes. For this reason, it's not possible to write a model generator that creates models for GraphQL entrypoints using DSL.

[More details on issue + current workarounds](https://docs.google.com/document/d/1eA4Q5-TeAKLoG540qvIy6_4Qz5vwIh3hF9h-1ohacP0/edit?usp=sharing)
[More details on implementation](https://docs.google.com/document/d/1L7b5zb1E3F4t0dDeWmvhfBp9KWeYTvW4ogjbu3YUnXc/edit?usp=sharing)

# This diff
Is the implementation. The main steps were:

1. Running Hack so it does not discard attributes
2. Adding a new heap to store inherited attribute info
3. Computing and storing attributes that the parents' of each method have when folded decls are constructed
4. Retrieving and using inherited attributes during DSL matching on attributes

Reviewed By: fahndrich

Differential Revision: D47609841

fbshipit-source-id: 24bf4f065bda33446c9e96c87125d03cf88c3d07