Fix utils test
Summary:
LLDB can't find the `HPHP::VMFixedVector<ObjectProps::quick_index>` type when running HHVM in buck. However, we can find it when running LLDB on HHVM directly (which is what I think we really care about anyway).

This test isn't about that type in particular, but rather it being able to get the type without template arguments, so let's just change the type to something that we can consistently find in both ways of running it.

Reviewed By: jtwarren

Differential Revision: D48359170

fbshipit-source-id: d10319be5963a556720e65812ebbaf134cc382dc