Various minor improvements to hhvm's lldb scripts, unit tests
Summary:
We don't need to launch the process for some of the tests. We can also be more accurate when looking up types by searching the first module (which I assume normally pertains to the hhvm executable) before searching the entire target if the first fails.

These changes might speed things up a little.

Reviewed By: clayborg

Differential Revision: D48365321

fbshipit-source-id: f186210195c78cea9ad6677274fb272e17ee84f9