plumb stream_file, hint_suffixes hh_client -> serverRpc
Summary:
This diff adds support for a new command, `--ide-find-refs-by-symbol3`
```
hh --ide-find-refs-by-symbol <symbol|action>
hh --ide-find-refs-by-symbol3 <symbol|action> <streamfile> <hints>
```
It plumbs it through as far as ServerRpc.

The intent of this is to eventually support (1) clientIdeDaemon getting glean hints about which files to look in for references, (2) streaming output.

The format of <streamfile> and <hints> is sort of implied by the code in findRefsWireFormat.ml. I left it opaque to callers.

Reviewed By: mheiber

Differential Revision: D48091861

fbshipit-source-id: f9951b2d10fed60d4e973d30a0d79c0c81c773a1