plumb stream_file, hint_suffixes glean -> hh_client
Summary:
This diff continues my work towards streaming find-refs. In particular,
1. With this diff, the find-refs implementation in clientIdeDaemon will always query glean first, and get a list of "hint" filenames
2. If VSCode send a partialResults token with its find-refs request, then clientLsp will (1) create a streaming-find-refs-results file and place in it the results from open files, (2) pass this filename and the list of hints on to "hh --ide-find-refs-by-symbol3"

This diff doesn't yet do anything with the partial results that get written into the streaming file, nor clean it up. That will happen later.

This diff comes with a perf hit. We will now get an extra 20-50ms per find-refs, the time it takes for a glean request. I think that's small enough that it's not even worth measuring.

Reviewed By: mheiber

Differential Revision: D48091860

fbshipit-source-id: 84aacd464b2bd14c64945de9c9017eacb18503f3