Decouple sink from return type
Summary:
This diff completes decoupling of sinks from return types. A sink is not a Thrift data type and it is already accessible via `t_function::sink` so there is no reason to duplicate it in the return type. This brings the AST closer to the actual grammar. Instead of this coupling we should provide convenient primitives for return types in target languages. One of such primitives, `function:cpp_return_type` has been added earlier. This diff adds another one, `function:cpp_void?` which specifies whether the generated C++ function is void.

There are some changes to the Rust fixture output for interactions which don't matter because interactions are not yet supported there.

Reviewed By: avalonalex

Differential Revision: D48379178

fbshipit-source-id: db5e5d0f9bc9d9015cd4cdf3911c9eebbebc134e