Approximate Sets Abstraction
Summary:
# Context
As part of the work on case types I need to expand the current model for data types. Currently as implemented a data type consists of a set of tags, with each tag being distinct from one another. Thus I utilize the standard OCaml Set implementation.

Now that we want to support multiple types with the object tag in a case type, we need a more sophisticated model of a data type.

# Solution
While experimenting on possible solutions, I determined the best way to add support for multiple objects was to come up with a hand-written set implementation designed specifically to answer if two sets are disjoint or not. The set implementation combines sets of atomic elements that can be compared using a relation operator. This operator determines if atomic elements are equal, subsets, disjoint or unknown.

This abstraction matches nicely with the model of multiple objects because we can take class tags as atomic elements and determine their relationship in this way. The unknown determination is particularly useful for reasoning about open classes and interfaces in a logical way.

To make testing this implementation easier, I've abstracted it into its own module that I can unit test on its own, thus this diff that introduces `ApproxSet`. The set abstraction is approximate because it is possible to construct sets where we cannot definitively determine if they are disjoint or not. In these cases we will conservatively determine the sets are not disjoint. The other reason the set is considered approximate is because atomic elements relationship may be unknown (i.e. in the case of two unrelated interfaces).

The bulk of the logic is in the `disjoint` function which applies a series of simplification rules taken from this [List of set identifies and relations](https://en.wikipedia.org/wiki/List_of_set_identities_and_relations).

Reviewed By: enetsee

Differential Revision: D48088684

fbshipit-source-id: cd8111386637af999ff91e9d19265a9d36644433