Fix handling of type aliases which resolve to "nothing"
Summary:
The "nothing" type-hint isn't valid for properties. If you
try to use one, the parser will recognize it and emit a fatal
instead. However, you can subvert this checking by instead making a
type alias to "nothing" and using that as the property type.

The runtime didn't check for this scenario and thus you could end up
with a property with an effective type of "nothing". However, if HHBBC
resolved the type alias, you could end up with the type-hint being
"nothing" directly, which breaks invariants.

First change the runtime to match the parser behavior. If we resolve a
type alias to "nothing", we'll fatal just as if the parser detected
it. If HHBBC attempts to resolve a property's type-hint to "nothing",
it will leave the type-hint as unresolved (which means the property is
really a Bottom since it can never be used).

Reviewed By: jano

Differential Revision: D48424344

fbshipit-source-id: f6e4f9cf7b8ec6e1b26a20e066b3e57dbf052b51