Make --keep-going the default
Summary:
In production scenarios we always use `--keep-going`. It makes sense to make it the default to streamline things a bit. Instead there's an opposite option now `--fail-fast`.

NOTE: we still have the `--keep-going` option that does nothing for now and is there only for backward compat until we change the command lines in all other places.

Reviewed By: aorenste

Differential Revision: D48431008

fbshipit-source-id: 8255cebd10ff98346a495ae64b92b2a51d4cce87