Cancel ping probes on destruction
Summary:
Crash: https://www.internalfb.com/logview/proxygen_crashes/627d1bb43b908c8c9f5c5ccfbd75e72a

My intuition is that the timeout fires and the session is already destroyed, causing the segfault seen in the crash. I don't have a good way to verify this - maybe someone can work with me on a unit test to specifically exercise this case.

Otherwise we can take the approach of landing this and waiting for the olb crashes to go away (currently the JK controlling upstream ping probes is enabled on 0.05% of olb hosts).

Reviewed By: hanidamlaj

Differential Revision: D48415117

fbshipit-source-id: e045a6421d91aaa18519ecbf5105586e0741d84e