Check for breakpoints before entering translations
Summary: Insert code at the beginning of each translation checking if a breakpoint may be set on any part of the function. If yes, exit the translation and run in the interpreter in `dispatchBB` mode. An RDS handle is abound to a function before creating any translation for that function or before resolving any breakpoint to that function. Resolving breakpoint initializes the handle which is checked by the inserted instruction. Once set, the breakpoint check would be true for the entire duration of the request. The check is also inserted after every call as a return jumps in the middle of a translation. The checks are omitted for builtins.

Reviewed By: paulbiss

Differential Revision: D47688565

fbshipit-source-id: 0e5fe8d2a176ce3cea9b38e687d4c768929a4525