Fix redefinition check
Summary:
This is embarrassing but it turned out that we only check for redefinition of structs, exceptions and interactions and not other kinds of definitions. For example, if a Thrift developer has code like this

```
struct A {}
enum A {}
```

it will pass the Thrift compiler but fail miserably when they try to compile the generated code. This is bad developer experience so let's fix it and move the check to the AST construction phase where it belongs and where we already have local type resolution.

Reviewed By: iahs

Differential Revision: D48413479

fbshipit-source-id: 9bae0a922415ea2d37fbdef2f3cf10abc7a9c8e7