Revert D48413479: Multisect successfully blamed "D48413479: [Thrift] Fix redefinition check" for test or build failures
Summary:
This diff is reverting D48413479
D48413479: [Thrift] Fix redefinition check by vitaut has been identified to be causing the following test or build failures:

Tests affected:
- [nodeapi/lib/fnc/privacylib/annotated_thrift:annotated_thrift_test - invalid_cases::invalid_cases::test_error_when_purpose_on_nested_field](https://www.internalfb.com/intern/test/844425005480132/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2834860
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: weirdev

Differential Revision: D48498902

fbshipit-source-id: 538ccf742f33bac7760ff893faad7e1342a3b3a4