Use uniform TODO style for hhbc and textual todos
Summary:
We used `TODO_<name>` for hhbc todos and `$todo.<name>` for textual todos. The latter was problematic because it triggered logic in infer where it would look for the 0-th argument (self) and fail because it's not there.

With this diff we use `TODO_<name>` for both types of todos because it doesn't trigger the logic around self arg in infer and is uniform.

Reviewed By: ezgicicek

Differential Revision: D48470686

fbshipit-source-id: 2f7fc7008693f39036b1ea5cd384e8b3ab824c5f