Mark declared members in depgraph
Summary: For the new fanout algorithm for member, we need to mark in the depgraph whether a member is declared by a (shallow) class. This diff does the marking. In Typing_class, right after adding a dep edge to the current module, we add edges to the new `Declares` dep variant from each member of the class.

Differential Revision: D47471289

fbshipit-source-id: 370c0c056270ae8c03717094789bab5a685cff0f