option to create file with list of reindexed files
Summary:
When indexing an increment, the hack indexer goes through all files,  but only reindexes a subset. It computes and returns the list of reindexed files so that they can be excluded from the base db.

Currently, this list is returned in the stderr log. This diffs adds a proper option instead.

Reviewed By: malanka

Differential Revision: D48516172

fbshipit-source-id: a702e597b459e65b45a347f6376e83829b5440e7