Pairwise Check for Overlapping Data Types in Case Types
Summary:
# Context
Case types require that each variant type maps to a disjoint set of values. The ensure this we compute data types for each type that corresponds to the set of possible values that type can contain. The current way we check for disjointness is by iterating over the set of possible tags and checking which variant types' data type intersect with that tag. If there are multiple variant types associated with a given tag then an error is reported.

This works fine given we are able to list all possible tags, but once we support multiple object types in a case type we will no longer be able to iterate over all possible tags.

# Solution
This diff changes how we perform validation by checking if each variant types' data type are disjoint pairwise. For instance if we have the case type:
```
case type CT = T1 | T2 | T3 | Tn;
```

We will check:
```
check_disjoint(T1, T2)
check_disjoint(T1, T3)
check_disjoint(T1, Tn)
check_disjoint(T2, T3)
check_disjoint(T2, Tn)
check_disjoint(T3, Tn)
```

If any pair is not disjoint an error is reported. This check is O(n^2) where n is the number of variant types in a case type, while the previous algorithm was O(t * n), where t is the number of potential tags. Assuming we keep `n` small (<16? <32?) the algorithmic performance difference shouldn't matter much, especially since we only need to check this when a case type is declared and not when it is used.

The other side effect of this change is the previous approach naturally would group common errors together, so we won't emit at most `t` errors. Because we are checking pairwise we will emit O(n^2) errors in the worst case. I experimented with methodologies for grouping related errors together to reduce the potential blow up, but it did make the validation code more difficult to reason about. In practice I do not think this will be a problem since I imagine in most case overlaps will most often happen between only two variants. If this becomes a source of user pain we can always implement the grouping logic.

Reviewed By: vassilmladenov

Differential Revision: D48419047

fbshipit-source-id: 4ba5c4ae4e77d15e76aec84cd2da0dbc6cc89f2b