Revert: Require Case Type Variants to be Recoverable
Summary:
{D44945717} was my first model for case types to ensure that we don't allow overlapping type declarations. Since then I've grown to prefer the analysis focused on data types and modeling the tags carried at runtime in Typing_case_types.DataType. Particularly the approach taken in D44945717 leads to difficult to comprehend error message, while the data type cased approach feels more actionable.

We are also moving pass the original idea that case types limits it to unions that we can refine. Now that we have specific refinement logic involving case types, there are conclusions we can arrive at for a case type that we wouldn't be able to arrive at for union type (taking advantage of the disjointness of runtime tags).

Based on all the above I rather revert the changes.

Note this is a manual revert since some relevant code changed since the original diff landed.

Reviewed By: vassilmladenov

Differential Revision: D48378787

fbshipit-source-id: b49c806402c3b581b3e889f79f0fd73fd7b68b46