Reprocess blocks when finishing an AddElem chain in HHBBC
Summary:
When finishing an AddElem chain, we push a type representing
a non-static version of the final array, but we simultaneously
const-prop away all the AddElem bytecodes. This means if we re-interp
the function (with no other changes), we'll get a better type. This
can break various invariants, because types shouldn't improve once we
reach a fixed-point. Fix this by forcing a block reprocess, to infer
the types of from the constprop.

Reviewed By: mdko

Differential Revision: D48547587

fbshipit-source-id: 9b92d81ab6adf6518bbbf718305a552711b6f11f