getWrappingTransport
Summary:
cristianlumezanu has changed `folly::AsyncSocket` so that observer functionality is specific to `AsyncSocket` and no longer part of `AsyncTransport`. This means that observers get passed an `AsyncSocket` object.

In some cases, an observer may want to be able to get information from a wrapping `AsyncTransport`. Using `derived_cast` will be insufficient in cases where `DecoratedAsyncTransportWrapper` is used, as `DecoratedAsyncTransportWrapper` holds a `unique_ptr` to the decorated transport instead of inheriting from it. To handle such instances, we store a pointer to the decorating transport in the decorated transport.

Reviewed By: bschlinker

Differential Revision: D47559814

fbshipit-source-id: 682fff8f6b199a1cc9636f4b4bdacc927d34ec12