Easy List.length = 0 -> List.is_empty optimizations
Summary:
I was talking with a personal friend learning OCaml yesterday and noticed that List has these neat compare_lengths and compare_length_with functions.

In our codebases, we do List.length on two lists a lot and compare the result, or we do it once and compare the result to a constant integer. Not only is this linear time, it also always traverses the entire list, when we could have bailed early (say we were checking that the length is greater than 2, e.g., a list of length 1000 still gets traversed all the way through).

 ---

All of that to say, that's not what this diff does , this diff just does the easy replacements of List.length <> 0 or similar to call Core.List.is_empty, which is essentially a pointer compare with the empty list and is faster in all cases than calling length then comparing to 0.

I think a follow up (that I won't do, but if anyone is interested) is to make utility functions List.shorter_than/longer_than/equal_lengths that wrap Caml.List.compare_lengths and Caml.List.compare_length_with in a nicer looking API, because they're pretty ugly to use directly, then migrate all `List.length x < int` and `List.length a < List.length b` calls or similar to it, if the length is only used in the comparison and is not subsequently used in the conditional branches.

On callsites where large lists are often passed, or lists of very disparate sizes are passed, it could be a potentially-significant perf win.

Reviewed By: alexkassil

Differential Revision: D48542763

fbshipit-source-id: b846bc32523b7b1b938cc033eb001d656970fa0f