Peek local's type when processing UnsetL in HHBBC
Summary:
If we're going to reduce away an UnsetL, we don't want to
mark the local as being read (because it won't be). Use peekLocRaw
instead of locRaw to avoid marking the local as being read. If we
don't reduce it, the logic below will mark it as being ready anyways.

Reviewed By: paulbiss

Differential Revision: D48570492

fbshipit-source-id: e1380cb035738a6cbb09288328b69051736833fc