Restore old String() method functionality
Summary:
`String()` method implementation between the new and legacy Thrift generators was not kept backwards compatible.
In the new generator, we opted in for a simpler Go-syntax representation.

As Nathan pointed out in D47303074 - the legacy representation actually looks better (more informative, easier to read).
It is also required for some existing unittests to migrate seamlessly.

In this diff, we are restoring the [legacy string representation](https://www.internalfb.com/code/fbsource/[ced2fba7370c]/fbcode/thrift/compiler/generate/t_go_generator.cc?lines=1589-1623). And we will be keeping it in the new generator going forward.

Reviewed By: abulimov

Differential Revision: D48286641

fbshipit-source-id: 47ee92f026837d83c6e7a638e04d5c852950cb3c