Simplify upper bound representation
Summary:
Do not try to reconstruct higher level concepts not tracked by HHVM, such as
upper bounds on a type variable defined on a function or class. This
methodology was fragile and not compatible with resolution certain aspects
such as nullability at parse time, or future needs that require multiple type
constraints on the same type info.

Instead, include all upper bounds explicitly as part of the type info.

Reviewed By: ricklavoie

Differential Revision: D48619545

fbshipit-source-id: ca31e92a737419a3fa6acda15a7c08df6e784cb7