tweak ServerFiles.path_of_root
Summary:
There's no urgency to this change, no goal.

I just looked at the code and thought "it's a fairly common convention that when you pass a filename and an extension then the kinds of extension you pass are '.txt' or '.jpg' or similar, including the dot". This allows for flexibility e.g. in cases where you don't want any extension at all. So making the change here.

Reviewed By: nt591

Differential Revision: D48711604

fbshipit-source-id: 263693d2277ebf881eed2728eaaa9ad6f7ee2cc4