Mitigate segment violation crash on startup/shutdown race
Summary:
When rev proxy fails to start mid-way and starts exiting, we sometimes crash with a segment violation. I am not exactly sure whether this is because the initialization has not quite finished or shutdown is racy, calling reset() of the connection manager before it finished draining.
In both cases the mitigation is the same, check the connection manager unique pointer before dereferencing it.

Context: https://fb.workplace.com/groups/715931878455430/posts/25095209603434318

Reviewed By: HsiehYuho

Differential Revision: D48666094

fbshipit-source-id: d8d8a86aff1ea2a8497071a8403d9eaab43818ab