Add Logging for completion_resolve and completion_resolve_location
Summary:
The code in clientIdeDaemon forks with completion_resolve and completion_resolve_location.
Today, we aren't sure which/why one of these gets called over the other.
In this diff, I'm adding very simple logging to both paths.
I wonder if maybe one of them is not called and then I can remove the dead path.

Reviewed By: ljw1004

Differential Revision: D48740067

fbshipit-source-id: a0f8bc286afe99a554038e95620adc9fcf108338