Depgraph: allow for multiple concurrent readers by using RwLock instead of Mutex
Summary:
This should allow better perf when computing fanouts in parallel.
(we currently only do that for non-class fanouts, though we might want to do it for classes in a near future).

It also allows to worry less about careful locking for performance and deadlocks (since we have two locks, we could in theory cause deadlocks if trying to lock both), and therefore simplify and clean up the code in a future diff.

Differential Revision: D48605477

fbshipit-source-id: 4ecf841c779ede00a8a7b2435b96341c71d502ca