Fix Ubuntu CI (#9391)
Summary:
Not all exists command have a alternative, but if there is one, a soft link will in the `/etc/alternatives`.
If you try to delete a alternative that not exists, will get this:`update-alternatives: error: no alternatives for cc`.
This PR added a condition when deleting the alternative, and the error not raised:
![image](https://github.com/facebook/hhvm/assets/120731947/473b927e-00dd-4120-bcb2-f2e5f7ed7526)

Footnote: The failing of step `Installing HHVM deps and building HHVM` is unrelated to this PR.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9391

Reviewed By: Wilfred

Differential Revision: D48672456

fbshipit-source-id: d58778b0ac32ac9e28506bff67f7951d5e96d94b