SDT function calling all-dynamic fall-back
Summary:
Our SDT function call logic is currently as follows.

* For each parameter, first attempt to check the actual argument type `t` against the precise parameter type `u` i.e. try `t <: u`. (Call this the STATIC check.) If this succeeds, move onto the next parameter.
* If it fails, check the actual argument type against the *pessimised* parameter type (i.e. try `t <: ~u`). (Call this the DYNAMIC check.)

If any argument needs a DYNAMIC check, we add `~` to the return type of the function.

A problem with this approach is that it may reject a call for which STATIC succeeds on the first argument then both STATIC and DYNAMIC fail on the second, but which would have succeeded with DYNAMIC then DYNAMIC. To address this, we use `Errors.do` to retry the whole call logic if the parameter-by-parameter check fails and we didn't already check every parameter using DYNAMIC.

Differential Revision: D48406781

fbshipit-source-id: 0731d3fe19f8d7afc418fbff7651b7f49ad726c2